requirements.txt
Django==1.6.2
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
youtube-dl==2014.02.28
newrelic==2.14.0.11
django-uuidfield==0.5.0
